make test more clear
