+++ schema-builder-library/src/internal_integration_tests/mod.rs
@@ -3,11 +3,11 @@ mod consts;
 
 #[cfg(feature = "integration")]
 #[cfg(test)]
-mod derive_schema_dotted_fields;
+mod derive_schema_heterogeneous;
 
 #[cfg(feature = "integration")]
 #[cfg(test)]
-mod derive_schema_heterogeneous;
+mod derive_schema_unmatchable_documents;
 
 #[cfg(feature = "integration")]
 #[cfg(test)]
